LibraryDone

+++ library/back/index.js
@@ -14,7 +14,7 @@ const connection = mysql.createConnection({
 });
 
 const app = express();
-const port = 3001;
+const port = 80;
 
 app.use(
 	cors({
@@ -109,9 +109,7 @@ const doAuth = (req, res, next) => {
 
 	connection.query(sql, token, (err, results) => {
 		if (err) {
-			res
-				.status(500)
-				.json({ message: { type: 'danger', text: 'Authorization failed' } });
+			res.status(500).json({ message: 'Server error on Auth' });
 		} else {
 			if (results.length > 0) {
 				const user = results[0];
@@ -124,28 +122,291 @@ const doAuth = (req, res, next) => {
 
 app.use(doAuth);
 
+// FRONT OFFICE
+
+app.get('/', (req, res) => {
+	const sort = req.query.sort || '';
+	let sql;
+
+	if (sort === 'name_asc') {
+		sql = `
+		SELECT a.id, a.name, a.surname, b.id as book_id, b.title, b.pages, b.genre, h.id as hero_id, h.name as hero, good, h.url as heroUrl, b.url as bookUrl, b.rate
+		FROM authors as a 
+		LEFT JOIN books as b
+		on a.id = b.author_id
+		LEFT JOIN heroes as h
+		on b.id = h.book_id
+		ORDER BY a.surname, a.name`;
+	} else if (sort === 'name_desc') {
+		sql = `
+		SELECT a.id, a.name, a.surname, b.id as book_id, b.title, b.pages, b.genre, h.id as hero_id, h.name as hero, good, h.url as heroUrl, b.url as bookUrl, b.rate
+		FROM authors as a 
+		LEFT JOIN books as b
+		on a.id = b.author_id
+		LEFT JOIN heroes as h
+		on b.id = h.book_id
+		ORDER BY a.surname DESC, a.name DESC`;
+	} else {
+		sql = `
+		SELECT a.id, a.name, a.surname, b.id as book_id, b.title, b.pages, b.genre, h.id as hero_id, h.name as hero, good, h.url as heroUrl, b.url as bookUrl, b.rate
+		FROM authors as a 
+		LEFT JOIN books as b
+		on a.id = b.author_id
+		LEFT JOIN heroes as h
+		on b.id = h.book_id`;
+	}
+	connection.query(sql, (err, results) => {
+		if (err) {
+			res.status(500).send(err);
+		} else {
+			res.json(results);
+		}
+	});
+});
+
+app.get('/hero/:slug', (req, res) => {
+	const sql = `
+	SELECT title, h.id, h.name, a.name as authorName, a.surname as authorSurname, good, title, book_id, image, h.url as heroUrl, b.url as bookUrl
+	FROM heroes as h
+	LEFT JOIN books as b
+	ON h.book_id = b.id
+	LEFT JOIN authors as a
+	ON b.author_id = a.id
+	WHERE h.url = ?
+	`;
+	connection.query(sql, [req.params.slug], (err, results) => {
+		if (err) {
+			res.status(500).send(err);
+		} else {
+			res.json(results[0]);
+		}
+	});
+});
+
+app.get('/book/:slug', (req, res) => {
+	const sql = `
+		SELECT b.id, title, pages, genre, a.name, surname, author_id, b.url AS bookUrl, h.url AS heroUrl, h.id AS hero_id, h.name AS hero, good, image
+		FROM books AS b
+		LEFT JOIN authors AS a
+		ON b.author_id = a.id
+		LEFT JOIN heroes AS h
+		ON b.id = h.book_id
+		WHERE b.url = ?
+	`;
+	connection.query(sql, [req.params.slug], (err, result) => {
+		if (err) {
+			res.status(500).send(err);
+		} else {
+			res.json(result);
+		}
+	});
+});
+
+app.get('/heroes-list', (req, res) => {
+	let sql;
+	let query;
+	const inPage = 2;
+	const page = req.query.page || 1;
+	let filter = req.query.filter || '';
+	let sort = req.query.sort || '';
+	filter = filter === 'good' ? 1 : filter === 'bad' ? 0 : '';
+	let total = 0;
+	let totalPages = 0;
+
+	if (filter === '') {
+		sql = `SELECT COUNT(*) AS total 
+		FROM heroes`;
+		query = [];
+	} else {
+		sql = `
+		SELECT COUNT(*) AS total 
+		FROM heroes
+		WHERE good = ? `;
+		query = [filter];
+	}
+	connection.query(sql, query, (err, countResult) => {
+		if (err) {
+			res.status(500).send(err);
+		} else {
+			total = countResult[0].total;
+			totalPages = Math.ceil(total / inPage);
+		}
+	});
+
+	if (filter === '' && sort === '') {
+		sql = `
+		SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+		LIMIT ?, ?
+	    `;
+		query = [(page - 1) * inPage, inPage];
+	} else if (filter !== '' && sort === '') {
+		sql = `
+		SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+		WHERE good = ?
+		LIMIT ?, ? `;
+		query = [filter, (page - 1) * inPage, inPage];
+	} else if (filter === '' && sort !== '') {
+		if (sort === 'name_asc') {
+			sql = `
+			SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+			ORDER BY name
+			LIMIT ?, ?
+			`;
+		} else if (sort === 'name_desc') {
+			sql = `
+			SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+			ORDER BY name DESC
+			LIMIT ?, ?
+			`;
+		} else {
+			sql = `
+			SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+			LIMIT ?, ?
+			`;
+		}
+		query = [(page - 1) * inPage, inPage];
+	} else {
+		if (sort === 'name_asc') {
+			sql = `
+			SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+			WHERE good = ?
+			ORDER BY name
+			LIMIT ?, ?
+			`;
+		} else if (sort === 'name_desc') {
+			sql = `
+			SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+			WHERE good = ?
+			ORDER BY name DESC
+			LIMIT ?,?
+			`;
+		} else {
+			sql = `
+			SELECT h.id, h.name, h.good, h.book_id, h.image, b.url, b.title
+		FROM heroes AS h
+		LEFT JOIN books AS b
+		ON h.book_id = b.id
+			WHERE good = ?
+			LIMIT ?, ?
+			`;
+		}
+		query = [filter, (page - 1) * inPage, inPage];
+	}
+	connection.query(sql, query, (err, result) => {
+		if (err) {
+			res.status(500).send(err);
+		} else {
+			res.json({ result, total, totalPages, page: +page });
+		}
+	});
+});
+
+app.get('/rating/:slug/:mark', (req, res) => {
+	setTimeout(_ => {
+		// fake delay
+
+		const sql = `
+	SELECT ratings
+	FROM books
+	WHERE url = ?
+	`;
+		connection.query(sql, [req.params.slug], (err, result) => {
+			if (err) {
+				res.status(500).send(err);
+			} else {
+				const ratings = JSON.parse(result[0].ratings);
+				const userMark = ratings.find(item => item.mark === req.params.mark);
+				const votes = ratings.length;
+				const sum = ratings.reduce((acc, item) => acc + +item.rate, 0);
+				const rate = +(sum / votes).toFixed(1);
+				res.json({
+					canRate: userMark === undefined ? true : false,
+					rate,
+					votes,
+					sum,
+				});
+			}
+		});
+	}, 2000);
+});
+
+app.post('/rating/:id/:mark', (req, res) => {
+	const { rate } = req.body;
+	const mark = req.params.mark;
+	const sql = `
+	SELECT ratings
+	FROM books
+	WHERE id = ?
+	`;
+	connection.query(sql, [req.params.id], (err, result) => {
+		if (err) {
+			res.status(500).send(err);
+		} else {
+			const ratings = JSON.parse(result[0].ratings);
+			ratings.push({ mark, rate });
+			const newRatings = JSON.stringify(ratings);
+			const votes = ratings.length;
+			const sum = ratings.reduce((acc, item) => acc + +item.rate, 0);
+			const newRate = +(sum / votes).toFixed(1);
+
+			const sql = `
+		UPDATE books
+		SET ratings = ?, rate = ?
+		WHERE id = ?
+		`;
+			connection.query(sql, [newRatings, newRate, req.params.id], err => {
+				if (err) {
+					res.status(500).send(err);
+				} else {
+					res.json({ success: true });
+				}
+			});
+		}
+	});
+});
+
 //login
 app.post('/login', (req, res) => {
 	const { username, password } = req.body;
 	const sql = 'SELECT * FROM users WHERE name = ? AND PASSWORD = ?';
 	connection.query(sql, [username, md5(password)], (err, results) => {
 		if (err) {
-			res.status(500).json({ message: { type: 'danger', text: 'Login failed' } });
+			res.status(500).json({ message: 'Server error On Login' });
 		} else {
 			if (results.length > 0) {
 				const token = md5(uuidv4());
 				const sql = 'UPDATE users SET session = ? WHERE id = ?';
 				connection.query(sql, [token, results[0].id], err => {
 					if (err) {
-						res
-							.status(500)
-							.json({ message: { type: 'danger', text: 'Login failed' } });
+						res.status(500).json({ message: 'Server error on Login' });
 					} else {
-						res.cookie('libSession', token, { maxAge: 1000 * 60 * 60 * 24 * 365 });
+						res.cookie('libSession', token, {
+							maxAge: 1000 * 60 * 60 * 24 * 365,
+							httpOnly: true,
+						});
 						res.json({
-							message: { type: 'success', text: 'Login succeeded' },
 							success: true,
-							token,
 							name: results[0].name,
 							role: results[0].role,
 							id: results[0].id,
@@ -153,9 +414,7 @@ app.post('/login', (req, res) => {
 					}
 				});
 			} else {
-				res
-					.status(401)
-					.json({ message: { type: 'danger', text: 'Invalid name or password' } });
+				res.status(401).json({ message: 'Invalid name or password' });
 			}
 		}
 	});
@@ -204,6 +463,10 @@ app.get('/stats', (req, res) => {
 });
 
 app.get('/authors', (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
+
 	const sql = 'SELECT * FROM authors';
 	connection.query(sql, (err, results) => {
 		if (err) {
@@ -215,6 +478,9 @@ app.get('/authors', (req, res) => {
 });
 
 app.get('/books', (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	const sql = `
 	SELECT b.id, title, pages, genre, name, surname, nickname, author_id
 	FROM books as b
@@ -230,14 +496,18 @@ app.get('/books', (req, res) => {
 });
 
 app.get('/heroes', (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	const sql = `
-	SELECT h.id, h.name, a.name AS authorName, a.surname AS authorSurname, good, title, book_id, h.image
-	FROM heroes as h
-	LEFT JOIN books as b
-	ON h.book_id = b.id
-	LEFT JOIN authors as a
-	ON b.author_id = a.id
-	ORDER BY h.id DESC`;
+    SELECT h.id, h.name, a.name AS authorName, a.surname AS authorSurname, good, title, book_id, h.image
+    FROM heroes as h
+    LEFT JOIN books as b 
+    ON h.book_id = b.id
+    LEFT JOIN authors as a
+    ON b.author_id = a.id
+    ORDER BY h.id DESC
+  `;
 	connection.query(sql, (err, results) => {
 		if (err) {
 			res.status(500).send(err);
@@ -250,6 +520,9 @@ app.get('/heroes', (req, res) => {
 app.post('/authors', (req, res) => {
 	// res.status(403).json({ status: 'login' });
 	// return;
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	const { name, surname, nickname, born } = req.body;
 	if (!name || !surname || !born) {
 		res.status(422).json({
@@ -257,9 +530,13 @@ app.post('/authors', (req, res) => {
 		});
 		return;
 	}
+
+	// const slug = title.toLowerCase().replace(/ /g, '-');
+	const slug = name.toLowerCase() + '-' + surname.toLowerCase();
+
 	const sql =
-		'INSERT INTO authors (name, surname, nickname, born) VALUES (?, ?, ?, ?)';
-	connection.query(sql, [name, surname, nickname, born], (err, result) => {
+		'INSERT INTO authors (name, surname, nickname, born, url) VALUES (?, ?, ?, ?, ?)';
+	connection.query(sql, [name, surname, nickname, born, slug], (err, result) => {
 		if (err) {
 			res.status(500).send(err);
 		} else {
@@ -274,6 +551,9 @@ app.post('/authors', (req, res) => {
 });
 
 app.post('/books', (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
 	const { title, pages, genre, author_id } = req.body;
@@ -286,9 +566,12 @@ app.post('/books', (req, res) => {
 		});
 		return;
 	}
+
+	const slug = title.toLowerCase().replace(/ /g, '-');
+
 	const sql =
-		'INSERT INTO books (title, pages, genre, author_id) VALUES (?, ?, ?, ?)';
-	connection.query(sql, [title, pages, genre, author_id], (err, result) => {
+		'INSERT INTO books (title, pages, genre, author_id, url) VALUES (?, ?, ?, ?, ?)';
+	connection.query(sql, [title, pages, genre, author_id, slug], (err, result) => {
 		if (err) {
 			res.status(500).send(err);
 		} else {
@@ -303,6 +586,9 @@ app.post('/books', (req, res) => {
 });
 
 app.post('/heroes', (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
 
@@ -317,10 +603,14 @@ app.post('/heroes', (req, res) => {
 		});
 		return;
 	}
-	const sql = 'INSERT INTO heroes (name, good, book_id, image) VALUES (?, ?, ?, ?)';
+
+	const slug = name.toLowerCase().replace(/ /g, '-');
+
+	const sql =
+		'INSERT INTO heroes (name, good, book_id, image, url) VALUES (?, ?, ?, ?, ?)';
 	connection.query(
 		sql,
-		[name, good, book_id, filename !== null ? '/images/' + filename : null],
+		[name, good, book_id, filename !== null ? '/images/' + filename : null, slug],
 		(err, result) => {
 			if (err) {
 				res.status(500).send(err);
@@ -337,6 +627,9 @@ app.post('/heroes', (req, res) => {
 });
 
 app.put(`/authors/:id`, (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
 	const { name, surname, nickname, born } = req.body;
@@ -346,22 +639,32 @@ app.put(`/authors/:id`, (req, res) => {
 		});
 		return;
 	}
+
+	const slug = name.toLowerCase() + '-' + surname.toLowerCase();
+
 	const sql =
-		'UPDATE authors SET name = ?, surname = ?, nickname = ?, born = ? WHERE id = ?';
-	connection.query(sql, [name, surname, nickname, born, req.params.id], err => {
-		if (err) {
-			res.status(500).send(err);
-		} else {
-			res.json({
-				success: true,
-				id: +req.params.id,
-				message: { type: 'info', text: 'Nice! Author updated!' },
-			});
+		'UPDATE authors SET name = ?, surname = ?, nickname = ?, born = ?, url = ? WHERE id = ?';
+	connection.query(
+		sql,
+		[name, surname, nickname, born, slug, req.params.id],
+		err => {
+			if (err) {
+				res.status(500).send(err);
+			} else {
+				res.json({
+					success: true,
+					id: +req.params.id,
+					message: { type: 'info', text: 'Nice! Author updated!' },
+				});
+			}
 		}
-	});
+	);
 });
 
 app.put(`/books/:id`, (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
 	const { title, pages, genre, author_id } = req.body;
@@ -374,22 +677,32 @@ app.put(`/books/:id`, (req, res) => {
 		});
 		return;
 	}
+
+	const slug = title.toLowerCase().replace(/ /g, '-');
+
 	const sql =
-		'UPDATE books SET title = ?, pages = ?, genre = ?, author_id = ? WHERE id = ?';
-	connection.query(sql, [title, pages, genre, author_id, req.params.id], err => {
-		if (err) {
-			res.status(500).send(err);
-		} else {
-			res.json({
-				success: true,
-				id: +req.params.id,
-				message: { type: 'info', text: 'The book has been updated!' },
-			});
+		'UPDATE books SET title = ?, pages = ?, genre = ?, author_id = ?, url = ? WHERE id = ?';
+	connection.query(
+		sql,
+		[title, pages, genre, author_id, slug, req.params.id],
+		err => {
+			if (err) {
+				res.status(500).send(err);
+			} else {
+				res.json({
+					success: true,
+					id: +req.params.id,
+					message: { type: 'info', text: 'The book has been updated!' },
+				});
+			}
 		}
-	});
+	);
 });
 
 app.put(`/heroes/:id`, (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
 	if (req.body.del) {
@@ -404,20 +717,25 @@ app.put(`/heroes/:id`, (req, res) => {
 		});
 		return;
 	}
+
+	const slug = name.toLowerCase().replace(/ /g, '-');
+
 	let sql;
 	let params;
-	if (req.body.del) {
-		sql = 'UPDATE heroes SET name = ?, good = ?, book_id = ?, image = ? WHERE id =?';
+	if (req.body.del || filename !== null) {
+		sql =
+			'UPDATE heroes SET name = ?, good = ?, book_id = ?, image = ?, url = ? WHERE id =?';
 		params = [
 			name,
 			good,
 			book_id,
 			filename !== null ? '/images/' + filename : null,
+			slug,
 			req.params.id,
 		];
 	} else {
-		sql = 'UPDATE heroes SET name = ?, good = ?, book_id = ? WHERE id = ?';
-		params = [name, good, book_id, req.params.id];
+		sql = 'UPDATE heroes SET name = ?, good = ?, book_id = ?, url = ? WHERE id = ?';
+		params = [name, good, book_id, slug, req.params.id];
 	}
 	connection.query(sql, params, err => {
 		if (err) {
@@ -434,6 +752,9 @@ app.put(`/heroes/:id`, (req, res) => {
 });
 
 app.delete(`/authors/:id`, (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
 	const sql = 'DELETE FROM authors WHERE id = ?';
@@ -461,21 +782,12 @@ app.delete(`/authors/:id`, (req, res) => {
 });
 
 app.delete(`/books/:id`, (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
 	let sql;
-	sql = 'SELECT image FROM heroes WHERE book_id = ?';
-	connection.query(sql, [req.params.id], (err, results) => {
-		if (err) {
-			res.status(500).send(err);
-		} else {
-			results.forEach(hero => {
-				if (hero.image) {
-					fs.unlinkSync('public/' + hero.image);
-				}
-			});
-		}
-	});
 
 	sql = 'DELETE FROM books WHERE id = ?';
 	connection.query(sql, [req.params.id], err => {
@@ -490,11 +802,26 @@ app.delete(`/books/:id`, (req, res) => {
 					text: 'Bum! Book deleted! All heroes from this book gone!',
 				},
 			});
+			sql = 'SELECT image FROM heroes WHERE book_id = ?';
+			connection.query(sql, [req.params.id], (err, results) => {
+				if (err) {
+					res.status(500).send(err);
+				} else {
+					results.forEach(hero => {
+						if (hero.image) {
+							fs.unlinkSync('public/' + hero.image);
+						}
+					});
+				}
+			});
 		}
 	});
 });
 
 app.delete(`/heroes/:id`, (req, res) => {
+	if (!checkUserisAuthorized(req.user, res, ['admin', 'user'])) {
+		return;
+	}
 	// res.status(403).json({ status: 'login' });
 	// return;
